Fix blank input bug
The query ternary statement only checks against `undefined`. But a blank text input in the browser doesn't have a value of undefined- this gets passed a blank string, which breaks the api call. 

With this change, if the query were *any* false-y value `''`/`undefined` etc etc, you'll get that curated query.